requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.11
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.5.2
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
crispy-bootstrap4==2024.10
dataclasses-json==0.6.7
Deprecated==1.2.15
dirtyjson==1.0.8
distro==1.9.0
Django==4.2.7
django-crispy-forms==2.3
djangorestframework==3.15.2
exceptiongroup==1.2.2
filetype==1.2.0
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.27.2
idna==3.10
jiter==0.7.1
joblib==1.4.2
llama-index-core==0.11.23
llama-index-readers-file==0.3.0
llama-parse==0.5.14
marshmallow==3.22.0
multidict==6.1.0
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.1
nltk==3.9.1
numpy==1.24.4
openai==1.55.0
packaging==24.2
pandas==2.0.3
pillow==10.4.0
pkg_resources==0.0.0
propcache==0.2.0
pydantic==2.10.0
pydantic_core==2.27.0
pypdf==5.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
reportlab==4.2.5
requests==2.32.3
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
sqlparse==0.5.2
striprtf==0.0.26
tenacity==8.5.0
threadpoolctl==3.5.0
tiktoken==0.7.0
tqdm==4.67.0
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
wrapt==1.16.0
yarl==1.15.2
